World Status Report 18 Apr 2020
This is an update on the numbers, but it is also a detailed, if somewhat haphazard description of some of the global problems of tracking and responding and correcting this Covid situation.

I extended my ideas about differentiating the polynomial (or function) that represents the cumulative cases to date. When that is differentiated, it should match a one lower order equation for the daily new cases, the cases per day, the case velocity.  The numbers are kind of lining up. At least the coefficients have the correct signs and magnitudes.

A global site to collect all the Covid data (globally neutral) could be set up at fairly low cost. I ask for someone with deep enough pockets to do that and do it right.  I suggested creating apps so that all the people in the world can report directly to the central site.

I am so tired, perhaps I should not even try.  But this is useless for me to do if I do not at least try to share it.  You are supposed to go do your part without supervision -- think on your own what needs to be done and start doing it.  Will it be wasteful and inefficient?  We can only try to do our best.

Richard Collins, Director, The Internet Foundation