Adding US Cumulative Fit and Number of Counties
It occurred to that the number of US counties should follow the same curve as the cases.  It turns out it is predictive and shows a peak two weeks earlier. This is the summary for the US cases and deaths. cases peak at WHO Situation Day 82.47 and the US Deaths at SitDay 88.09. Below is the number of counties and it peaked at SitDay 64.22. From first reports it is (18 days for cases and then 26 days for deaths.  If you look at Bangladesh is is infected all over.  I think in another couple of weeks there will be more deaths and it will grow. They are very good but it might be hard to get enough.  You should ask them quietly if they need help, not tell them. No one in the WHO system answers email.

The cases follow activity and the age structure tells how many cases.  That can be calibrated to the cases. They cases also give the peaks.  I think i have most of the pieces, at least for a general picture.  They guys at Mathematica/Wolfram have done diffusion studies, but just looking at the roads and early reports tell the story. The cities all are sufficiently dense to support high rates of infection. Then about 18 days from first reports to serious cases. About one in ten infected has serious problems. The number of cases requiring hospitalization is about one tenth the infected population.  It varies but that is what I use for eyeballing a situation.

88.09 - 64.22 = 23.87  --- deaths after first reports
82.47 - 64.22 = 18.25  --- cases arriving for treatment after first reports
88.09 - 82.47 =  5.62  ---  deaths take longer on average.  Are the hospital stays normally distributed?  How long after they arrive do people leave the hospital.  Are there any datasets from countries?

This is NYTimes Counties and ECDC US Report through WHO for the US

Once people get this, it is probably a normal distribution on when symptoms occur.  I have case data that I got today. Will check that, but I should be able to run the counties.  It is midnight.  Looks like I don't get to sleep tonight.

Same with the middle east. And it might be hard to deal with it because they might not have enough food and supplies for a 6 week lockdown.  I am going to check all the countries as soon as I get the data together.  The cumulative normal works better than the logistic and is familiar to more people.  I have to write a solver, that I can do from memory from long ago. I am not looking for perfection I am looking for ways to grab the data and use it to make decisions quickly and mostly in the right direction.  Often "on time" is vastly more important than "exactly right".