Population by single year Dominican Republic Master no video
I updated the Master twice and posted for those dates, no video
The Dominican Republic is just a look at their population projects that can be used for the old method of projecting teacher and education costs.
The UN population by single year of age spreadsheets were hard to read and to use, so I cleaned them up.